#2003: extract telemetry counters + diagnostics into bpf_map/metrics.rs
Pure code motion, second step of the bpf_map/mod.rs split. Lifts the
telemetry cluster — items that surface map occupancy and publish health
to the CLI / Prometheus collectors but take no part in the forwarding
decision — into a new bpf_map/metrics.rs.

Moved verbatim (no signature or logic change): diagnose_raw_ring_state
(the getsockopt/mmap raw XDP-ring probe), count_bpf_session_entries and
dump_bpf_session_entries (USERSPACE_SESSIONS iteration diagnostics), and
the publish counters SESSION_PUBLISH_VERIFY_OK/FAIL,
SESSION_PUBLISH_ERRORS_SHARED, SESSION_CREATIONS_LOGGED, and the
debug-gated ICMPV6_EMBED_LOGGED.

Original pub(super) visibility (= pub(in crate::afxdp)) is preserved
exactly and re-exported with `pub(in crate::afxdp) use metrics::*;`, so
both bare-name call sites (via the afxdp glob) and the explicit
`crate::afxdp::bpf_map::SESSION_PUBLISH_ERRORS_SHARED` paths in
session_glue/forwarding resolve unchanged.

Validation: cargo build --release and cargo test --release both clean
with this intermediate state (ha + metrics split so far). The pin
cluster (OwnedFd + degraded-path reader) follows in the next commit.
