+++ userspace-dp/src/afxdp/bpf_map/mod.rs
@@ -11,81 +11,6 @@ pub(super) fn uses_kernel_local_session_map_entry(
         && decision.resolution.tunnel_endpoint_id == 0
 }
 
-pub(super) fn diagnose_raw_ring_state(
-    sock_fd: c_int,
-) -> Option<(u32, u32, u32, u32, u32, u32, u32, u32)> {
-    // SOL_XDP (283) comes from afxdp/mod.rs via `use super::*` —
-    // #1826 deduplicated the former local copy.
-    const XDP_MMAP_OFFSETS: i32 = 1;
-    const XDP_PGOFF_RX_RING: i64 = 0;
-    const XDP_PGOFF_TX_RING: i64 = 0x80000000;
-    const XDP_UMEM_PGOFF_FILL_RING: i64 = 0x100000000;
-    const XDP_UMEM_PGOFF_COMPLETION_RING: i64 = 0x180000000;
-
-    // xdp_mmap_offsets_v2 (kernel >= 5.4): 4 rings × 4 fields × u64 each
-    #[repr(C)]
-    #[derive(Default)]
-    struct XdpRingOffset {
-        producer: u64,
-        consumer: u64,
-        desc: u64,
-        flags: u64,
-    }
-    #[repr(C)]
-    #[derive(Default)]
-    struct XdpMmapOffsets {
-        rx: XdpRingOffset,
-        tx: XdpRingOffset,
-        fr: XdpRingOffset,
-        cr: XdpRingOffset,
-    }
-
-    let mut off = XdpMmapOffsets::default();
-    let mut optlen = core::mem::size_of::<XdpMmapOffsets>() as libc::socklen_t;
-    let rc = unsafe {
-        libc::getsockopt(
-            sock_fd,
-            SOL_XDP,
-            XDP_MMAP_OFFSETS,
-            (&mut off as *mut XdpMmapOffsets).cast::<libc::c_void>(),
-            &mut optlen,
-        )
-    };
-    if rc != 0 {
-        return None;
-    }
-
-    fn read_ring_pair(sock_fd: c_int, off: &XdpRingOffset, pgoff: i64) -> (u32, u32) {
-        let map_len = (off.desc.max(off.consumer).max(off.producer) + 8) as usize;
-        let mmap_ptr = unsafe {
-            libc::mmap(
-                core::ptr::null_mut(),
-                map_len,
-                libc::PROT_READ,
-                libc::MAP_SHARED,
-                sock_fd,
-                pgoff,
-            )
-        };
-        if mmap_ptr == libc::MAP_FAILED {
-            return (0, 0);
-        }
-        let prod = unsafe { *(mmap_ptr.byte_add(off.producer as usize) as *const u32) };
-        let cons = unsafe { *(mmap_ptr.byte_add(off.consumer as usize) as *const u32) };
-        unsafe { libc::munmap(mmap_ptr, map_len) };
-        (prod, cons)
-    }
-
-    let (rx_prod, rx_cons) = read_ring_pair(sock_fd, &off.rx, XDP_PGOFF_RX_RING);
-    let (fr_prod, fr_cons) = read_ring_pair(sock_fd, &off.fr, XDP_UMEM_PGOFF_FILL_RING);
-    let (tx_prod, tx_cons) = read_ring_pair(sock_fd, &off.tx, XDP_PGOFF_TX_RING);
-    let (cr_prod, cr_cons) = read_ring_pair(sock_fd, &off.cr, XDP_UMEM_PGOFF_COMPLETION_RING);
-
-    Some((
-        rx_prod, rx_cons, fr_prod, fr_cons, tx_prod, tx_cons, cr_prod, cr_cons,
-    ))
-}
-
 pub(super) struct OwnedFd {
     pub(super) fd: c_int,
 }
@@ -608,160 +533,6 @@ pub(super) fn verify_session_key_in_bpf(map_fd: c_int, key: &SessionKey) -> bool
     rc == 0
 }
 
-/// Count total entries in the BPF USERSPACE_SESSIONS map.
-pub(super) fn count_bpf_session_entries(map_fd: c_int) -> u32 {
-    let mut count = 0u32;
-    let key_size = core::mem::size_of::<UserspaceSessionMapKey>();
-    let mut key = vec![0u8; key_size];
-    let mut next_key = vec![0u8; key_size];
-    // First key
-    let rc = unsafe {
-        libbpf_sys::bpf_map_get_next_key(
-            map_fd,
-            core::ptr::null(),
-            next_key.as_mut_ptr().cast::<c_void>(),
-        )
-    };
-    if rc != 0 {
-        return 0;
-    }
-    count += 1;
-    key.copy_from_slice(&next_key);
-    loop {
-        let rc = unsafe {
-            libbpf_sys::bpf_map_get_next_key(
-                map_fd,
-                key.as_ptr().cast::<c_void>(),
-                next_key.as_mut_ptr().cast::<c_void>(),
-            )
-        };
-        if rc != 0 {
-            break;
-        }
-        count += 1;
-        key.copy_from_slice(&next_key);
-        if count > 10000 {
-            break; // safety limit
-        }
-    }
-    count
-}
-
-/// Dump first N entries from the BPF USERSPACE_SESSIONS map for debugging.
-#[allow(unused_variables)]
-pub(super) fn dump_bpf_session_entries(map_fd: c_int, max_entries: u32) {
-    let key_size = core::mem::size_of::<UserspaceSessionMapKey>();
-    let mut key_bytes = vec![0u8; key_size];
-    let mut next_key_bytes = vec![0u8; key_size];
-    let mut value = 0u8;
-    let mut count = 0u32;
-    // First key
-    let rc = unsafe {
-        libbpf_sys::bpf_map_get_next_key(
-            map_fd,
-            core::ptr::null(),
-            next_key_bytes.as_mut_ptr().cast::<c_void>(),
-        )
-    };
-    if rc != 0 {
-        debug_log!("BPF_MAP_DUMP: empty (no entries)");
-        return;
-    }
-    loop {
-        // Read the key as UserspaceSessionMapKey
-        let map_key: UserspaceSessionMapKey =
-            unsafe { core::ptr::read(next_key_bytes.as_ptr().cast()) };
-        let _ = unsafe {
-            libbpf_sys::bpf_map_lookup_elem(
-                map_fd,
-                next_key_bytes.as_ptr().cast::<c_void>(),
-                (&mut value as *mut u8).cast::<c_void>(),
-            )
-        };
-        #[cfg(feature = "debug-log")]
-        {
-            let src_ip = if map_key.addr_family == libc::AF_INET as u8 {
-                format!(
-                    "{}.{}.{}.{}",
-                    map_key.src_addr[0],
-                    map_key.src_addr[1],
-                    map_key.src_addr[2],
-                    map_key.src_addr[3]
-                )
-            } else {
-                format!(
-                    "v6[{:02x}{:02x}::{:02x}{:02x}]",
-                    map_key.src_addr[0],
-                    map_key.src_addr[1],
-                    map_key.src_addr[14],
-                    map_key.src_addr[15]
-                )
-            };
-            let dst_ip = if map_key.addr_family == libc::AF_INET as u8 {
-                format!(
-                    "{}.{}.{}.{}",
-                    map_key.dst_addr[0],
-                    map_key.dst_addr[1],
-                    map_key.dst_addr[2],
-                    map_key.dst_addr[3]
-                )
-            } else {
-                format!(
-                    "v6[{:02x}{:02x}::{:02x}{:02x}]",
-                    map_key.dst_addr[0],
-                    map_key.dst_addr[1],
-                    map_key.dst_addr[14],
-                    map_key.dst_addr[15]
-                )
-            };
-            debug_log!(
-                "BPF_MAP_DUMP[{}]: af={} proto={} {}:{} -> {}:{} val={}",
-                count,
-                map_key.addr_family,
-                map_key.protocol,
-                src_ip,
-                map_key.src_port,
-                dst_ip,
-                map_key.dst_port,
-                value,
-            );
-        }
-        count += 1;
-        if count >= max_entries {
-            break;
-        }
-        key_bytes.copy_from_slice(&next_key_bytes);
-        let rc = unsafe {
-            libbpf_sys::bpf_map_get_next_key(
-                map_fd,
-                key_bytes.as_ptr().cast::<c_void>(),
-                next_key_bytes.as_mut_ptr().cast::<c_void>(),
-            )
-        };
-        if rc != 0 {
-            break;
-        }
-    }
-    debug_log!("BPF_MAP_DUMP: total={count} entries");
-}
-
-pub(super) static SESSION_PUBLISH_VERIFY_OK: AtomicU64 = AtomicU64::new(0);
-pub(super) static SESSION_PUBLISH_VERIFY_FAIL: AtomicU64 = AtomicU64::new(0);
-/// #1789: failed USERSPACE_SESSIONS BPF-map publishes from call sites
-/// that have no per-binding context (HA `upsert_synced_session`,
-/// session-glue worker publishes, post-reconcile `replay_synced_sessions`,
-/// activation/reverse prewarm). Always-on (NOT debug-gated, unlike the
-/// VERIFY counters above which only move under `debug-log`): a swallowed
-/// publish `Err` means the XDP shim never learns the key and the flow
-/// takes the NO_SESSION degraded path. Per-binding worker poll sites use
-/// `BindingLiveState::session_publish_errors` instead; the two are summed
-/// by `Coordinator::session_publish_errors_total()` and surfaced as
-/// `xpf_userspace_session_publish_errors_total`.
-pub(super) static SESSION_PUBLISH_ERRORS_SHARED: AtomicU64 = AtomicU64::new(0);
-pub(super) static SESSION_CREATIONS_LOGGED: AtomicU64 = AtomicU64::new(0);
-#[cfg(feature = "debug-log")]
-pub(super) static ICMPV6_EMBED_LOGGED: AtomicU32 = AtomicU32::new(0);
-
 // The pinned map path keeps the historical "fallback" spelling for
 // mixed-version shim compatibility. Operator-facing names use
 // degraded-path terminology.
@@ -943,12 +714,16 @@ mod tests;
 
 mod publish_conntrack;
 
-// #2003: behaviour-preserving code motion. The HA liveness-slot cluster
-// (XSK + heartbeat map writes that gate active-binding state) lives in
-// `ha.rs`; items keep their original `pub(in crate::afxdp)` visibility and
-// are re-exported so the parent `afxdp` glob (`use self::bpf_map::*`)
-// resolves them by bare name exactly as before. Mirrors the
-// `publish_conntrack.rs` precedent (#1356).
+// #2003: behaviour-preserving code motion. Each submodule owns one
+// cluster — `ha.rs` the HA liveness-slot writes (XSK + heartbeat map
+// updates that gate active-binding state), `metrics.rs` the telemetry
+// counters plus the raw-ring / session-map diagnostics. Items keep their
+// original `pub(in crate::afxdp)` visibility and are re-exported so the
+// parent `afxdp` glob (`use self::bpf_map::*`) and the relocated
+// `bpf_map_tests.rs` (`use super::*`) resolve them by bare name exactly as
+// before. Mirrors the `publish_conntrack.rs` precedent (#1356).
 mod ha;
+mod metrics;
 
 pub(in crate::afxdp) use ha::*;
+pub(in crate::afxdp) use metrics::*;
